+++ extensions/cronet/src/main/java/com/google/android/exoplayer2/ext/cronet/CronetEngineWrapper.java
@@ -73,25 +73,29 @@ public final class CronetEngineWrapper {
   public static final int SOURCE_UNAVAILABLE = 4;
 
   /**
-   * Creates a wrapper for a {@link CronetEngine} which automatically selects the most suitable
-   * {@link CronetProvider}. Sets wrapper to prefer natively bundled Cronet over GMSCore Cronet
-   * if both are available.
+   * Creates a wrapper for a {@link CronetEngine} built using the most suitable {@link
+   * CronetProvider}. When natively bundled Cronet and GMSCore Cronet are both available, the
+   * natively bundled provider is preferred.
    *
    * @param context A context.
    */
   public CronetEngineWrapper(Context context) {
-    this(context, false);
+    this(context, /* userAgent= */ null, /* preferGMSCoreCronet= */ false);
   }
 
   /**
-   * Creates a wrapper for a {@link CronetEngine} which automatically selects the most suitable
-   * {@link CronetProvider} based on user preference.
+   * Creates a wrapper for a {@link CronetEngine} built using the most suitable {@link
+   * CronetProvider}. When natively bundled Cronet and GMSCore Cronet are both available, {@code
+   * preferGMSCoreCronet} determines which is preferred.
    *
    * @param context A context.
+   * @param userAgent A default user agent, or {@code null} to use a default user agent of the
+   *     {@link CronetEngine}.
    * @param preferGMSCoreCronet Whether Cronet from GMSCore should be preferred over natively
    *     bundled Cronet if both are available.
    */
-  public CronetEngineWrapper(Context context, boolean preferGMSCoreCronet) {
+  public CronetEngineWrapper(
+      Context context, @Nullable String userAgent, boolean preferGMSCoreCronet) {
     CronetEngine cronetEngine = null;
     @CronetEngineSource int cronetEngineSource = SOURCE_UNAVAILABLE;
     List<CronetProvider> cronetProviders = new ArrayList<>(CronetProvider.getAllProviders(context));
@@ -108,7 +112,11 @@ public final class CronetEngineWrapper {
     for (int i = 0; i < cronetProviders.size() && cronetEngine == null; i++) {
       String providerName = cronetProviders.get(i).getName();
       try {
-        cronetEngine = cronetProviders.get(i).createBuilder().build();
+        CronetEngine.Builder cronetEngineBuilder = cronetProviders.get(i).createBuilder();
+        if (userAgent != null) {
+          cronetEngineBuilder.setUserAgent(userAgent);
+        }
+        cronetEngine = cronetEngineBuilder.build();
         if (providerComparator.isNativeProvider(providerName)) {
           cronetEngineSource = SOURCE_NATIVE;
         } else if (providerComparator.isGMSCoreProvider(providerName)) {
@@ -133,9 +141,9 @@ public final class CronetEngineWrapper {
   }
 
   /**
-   * Creates a wrapper for an existing CronetEngine.
+   * Creates a wrapper for an existing {@link CronetEngine}.
    *
-   * @param cronetEngine An existing CronetEngine.
+   * @param cronetEngine The CronetEngine to wrap.
    */
   public CronetEngineWrapper(CronetEngine cronetEngine) {
     this.cronetEngine = cronetEngine;