+++ extensions/okhttp/src/main/java/com/google/android/exoplayer2/ext/okhttp/OkHttpDataSourceFactory.java
@@ -15,7 +15,6 @@
  */
 package com.google.android.exoplayer2.ext.okhttp;
 
-import static com.google.android.exoplayer2.ExoPlayerLibraryInfo.DEFAULT_USER_AGENT;
 
 import androidx.annotation.Nullable;
 import com.google.android.exoplayer2.upstream.HttpDataSource;
@@ -40,7 +39,7 @@ public final class OkHttpDataSourceFactory extends BaseFactory {
    *     by the sources created by the factory.
    */
   public OkHttpDataSourceFactory(Call.Factory callFactory) {
-    this(callFactory, DEFAULT_USER_AGENT, /* listener= */ null, /* cacheControl= */ null);
+    this(callFactory, /* userAgent= */ null, /* listener= */ null, /* cacheControl= */ null);
   }
 
   /**